+++ r2-navigator/src/main/java/org/readium/r2/navigator/R2BasicWebView.kt
@@ -10,6 +10,7 @@
 package org.readium.r2.navigator
 
 import android.content.Context
+import android.content.SharedPreferences
 import android.graphics.PointF
 import android.os.Build
 import android.text.Html
@@ -26,8 +27,8 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.jsoup.Jsoup
 import org.jsoup.safety.Whitelist
-import org.readium.r2.shared.getAbsolute
 import org.readium.r2.shared.publication.ReadingProgression
+import org.readium.r2.shared.util.Href
 
 
 /**
@@ -38,6 +39,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
 
     lateinit var listener: Listener
     lateinit var navigator: Navigator
+    internal var preferences: SharedPreferences? = null
 
     var progression: Double = 0.0
     var overrideUrlLoading = true
@@ -138,7 +140,7 @@ open class R2BasicWebView(context: Context, attrs: AttributeSet) : WebView(conte
             val href = noteref.attr("href")
             if (href.indexOf("#") > 0) {
                 val id = href.substring(href.indexOf('#') + 1)
-                var absolute = getAbsolute(href, resourceUrl!!)
+                var absolute = Href(href, baseHref = resourceUrl!!).percentEncodedString
                 absolute = absolute.substring(0, absolute.indexOf("#"))
                 val document = Jsoup.connect(absolute).get()
                 val aside = document.select("aside#$id").first()?.html()