+++ r2-navigator/src/main/java/org/readium/r2/navigator/audio/AudioActivity.kt
@@ -11,11 +11,10 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Observer
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.readium.r2.navigator.MediaNavigator
 import org.readium.r2.navigator.R
-import org.readium.r2.navigator.media.*
+import org.readium.r2.navigator.media.MediaService
 import org.readium.r2.shared.AudioSupport
 import org.readium.r2.shared.FragmentNavigator
 import org.readium.r2.shared.extensions.getPublicationOrNull
@@ -55,7 +54,7 @@ open class AudioActivity : AppCompatActivity() {
         publicationId = mediaNavigator.publicationId
         mediaNavigator.play()
 
-        supportFragmentManager.fragmentFactory = AudioNavigatorFragment.Factory(mediaNavigator)
+        supportFragmentManager.fragmentFactory = AudioNavigatorFragment.createFactory(mediaNavigator)
 
         super.onCreate(savedInstanceState)